api/requirements.txt
flask
python-dotenv
psycopg2-binary
pyjwt
bcrypt
sentence_transformers